agent/requirements.txt
google-adk
mcp[cli]
fastapi
uvicorn[standard]
httpx
requests
